Add support for multiple client IDs

+++ tests/index.spec.ts
@@ -12,27 +12,48 @@ const { writeFile, unlink } = promises;
 const filename = resolve(__dirname, '..', 'public', 'index.html');
 
 test('find alerts', async ({ page }) => {
+	const credentials = JSON.parse(process.env.CREDENTIALS || '[]');
+
+	if (!credentials || !Array.isArray(credentials)) {
+		return;
+	}
+
+	const results: string[] = [];
+
 	if (existsSync(filename)) {
 		await unlink(filename);
 	}
 
-	await page.goto('https://tok.erpsever.bg/');
+	for (const credential of credentials) {
+		if (!Array.isArray(credentials) || credential.length !== 2) {
+			continue;
+		}
 
-	await page.locator('[data-customer-type="personal"]').click();
-	await page.locator('[name="username"]').fill(process.env.USERNAME || '');
-	await page.locator('[name="pin"]').fill(process.env.PASSWORD || '');
-	await page.locator('[type="submit"]').click();
+		const [username, password] = credential;
 
-	const table = page.locator('.table-responsive');
-	const results = await table.innerHTML();
+		await page.goto('https://tok.erpsever.bg/');
 
-	if (!results.includes('Няма прекъсвания за обекти на клиента')) {
-		await email(results);
+		await page.locator('[data-customer-type="personal"]').click();
+		await page.locator('[name="username"]').fill(username.toString());
+		await page.locator('[name="pin"]').fill(password.toString());
+		await page.locator('[type="submit"]').click();
+
+		const table = page.locator('.table-responsive');
+
+		results.push(await table.innerHTML());
+
+		await page.locator('[href="/logout"]').click();
+	}
+
+	const messages = results.filter(item => !item.includes('Няма прекъсвания за обекти на клиента'));
+
+	if (messages.length > 0) {
+		await email(messages.join('<br><br>'));
 	}
 
 	await writeFile(
 		filename,
-		minify(html(results), {
+		minify(html(results.join('<br><br>')), {
 			collapseWhitespace: true
 		}),
 		'utf-8'
